requirements.txt
fastmcp>=0.1.0
python-dotenv>=1.0.0
requests>=2.31.0
streamlit>=1.32.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
